LayoutTests/http/tests/loading/basic-auth-remove-credentials-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
127.0.0.1:8000 - didReceiveAuthenticationChallenge - Responding with first:first-pw
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didStartProvisionalLoadForFrame
127.0.0.1:8000 - didReceiveAuthenticationChallenge - Responding with second:second-pw
frame "<!--framePath //<!--frame1-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame1-->-->" - didFinishLoadForFrame
This test verifies that we are able to remove credentials, by checking that we get an authenticaton.
challenge for a path we normally would not get a challenge for, after having removed all credentials.
The first frame's path is /loading/resources/protected-resource.php, and we should get a challenge for it.
It will be authorized with first/first-pw.
The second frame's path is identical to the first frame, and we should get a challenge for it, because credentials have been deleted.
It will be authorized with second/second-pw.
  

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Authenticated as user: first password: first-pw

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
Authenticated as user: second password: second-pw
